requirements.txt
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.4
importlib-resources==5.12.0
joblib==1.2.0
kiwisolver==1.4.4
kneed==0.8.3
matplotlib==3.7.1
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
threadpoolctl==3.1.0
tzdata==2023.3
zipp==3.15.0

fastapi~=0.95.1

